method-prefix-refactor

+++ sagittarius/src/main/java/dev/samkist/lumae/sagittarius/data/models/contexts/TransactionContext.java
@@ -3,5 +3,5 @@ package dev.samkist.lumae.sagittarius.data.models.contexts;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-public record TransactionContext(String uuid, Long amount) {
+public record TransactionContext(String uuid, Double amount) {
 }
